Constants streamlined
