+++ src/main/java/com/darzee/shankh/constants/ImageLinks.java
@@ -1,0 +1,154 @@
+package com.darzee.shankh.constants;
+
+import org.springframework.beans.factory.annotation.Value;
+
+public class ImageLinks {
+
+    @Value("${amazonProperties.s3.staticBucketName}")
+    private static String staticBucketName;
+    public static final String DRESS_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress.svg";
+    public static final String DRESS_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress1_len.svg";
+    public static final String DRESS_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress2_shoulder.svg";
+    public static final String DRESS_UPPER_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress3_upperchest.svg";
+    public static final String DRESS_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress4_bust.svg";
+    public static final String DRESS_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress5_waist.svg";
+    public static final String DRESS_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress6_seat.svg";
+    public static final String DRESS_ARMHOLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress7_armhole.svg";
+    public static final String DRESS_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress8_slevlen.svg";
+    public static final String DRESS_SLEEVE_CIRCUMFERENCE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress9_slevcircum.svg";
+    public static final String DRESS_FRONT_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress10_frontneckdept.svg";
+    public static final String DRESS_BACK_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Dress/dress11_backneckdept.svg";
+    public static final String GOWN_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/nightgown.svg";
+
+    public static final String GOWN_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_length.svg";
+    public static final String GOWN_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_bust.svg";
+    public static final String GOWN_ARMHOLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_armhole.svg";
+    public static final String GOWN_UPPER_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_upperchest.svg";
+    public static final String GOWN_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_waist.svg";
+    public static final String GOWN_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_shoulder.svg";
+    public static final String GOWN_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_slevelen.svg";
+    public static final String GOWN_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_seat.svg";
+    public static final String GOWN_SLEEVE_CIRCUMFERENCE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_slevcircum.svg";
+    public static final String GOWN_FRONT_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_frontneck.svg";
+    public static final String GOWN_BACK_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_backneck.svg";
+    public static final String UNDER_SKIRT_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Underskirt/Underskirt.svg";
+
+    public static final String UNDER_SKIRT_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Underskirt/us_length.svg";
+    public static final String UNDER_SKIRT_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Underskirt/us_waist.svg";
+    public static final String BLOUSE_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse.svg";
+    public static final String BLOUSE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse1_len.svg";
+    public static final String BLOUSE_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse2_bust.svg";
+    public static final String BLOUSE_UPPOER_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse3_upperchest.svg";
+    public static final String BLOUSE_BELOW_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse4_belowbust.svg";
+    public static final String BLOUSE_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse5_shoulder.svg";
+    public static final String BLOUSE_ARMHOLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse6_armhole.svg";
+    public static final String BLOUSE_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse7_slevelen.svg";
+    public static final String BLOUSE_SLEEVE_CICUM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse8_slevcircumeference.svg";
+    public static final String BLOUSE_BICEP_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse9_bicep.svg";
+    public static final String BLOUSE_ELBOW_ROUND_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse10_elbowround.svg";
+    public static final String BLOUSE_APEX_TO_APEX_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse11_apextoapex.svg";
+    public static final String BLOUSE_SHOULDER_TO_APEX_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse12_shouldertoapex.svg";
+    public static final String BLOUSE_FRONT_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse13_frontneckdepth.svg";
+    public static final String BLOUSE_BACK_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse14_backneckdepth.svg";
+    public static final String BLOUSE_CROSS_FRONT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse15_crossfront.svg";
+    public static final String BLOUSE_CROSS_BACK_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Blouse/blouse16_crossback.svg";
+    public static final String SHIRT_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt.svg";
+    public static final String SHIRT_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_length.svg";
+    public static final String SHIRT_NECK_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_neck.svg";
+    public static final String SHIRT_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_shoulder.svg";
+    public static final String SHIRT_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_chest.svg";
+    public static final String SHIRT_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_waist.svg";
+    public static final String SHIRT_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_seat.svg";
+    public static final String SHIRT_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_sleeve.svg";
+    public static final String SHIRT_SLEEVE_CIRCUM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Shirt/shirt_sleevecircum.svg";
+    public static final String PANTS_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants.svg";
+    public static final String PANTS_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants1_waist.svg";
+    public static final String PANTS_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants2_seat.svg";
+    public static final String PANTS_THIGH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants3_thie.svg";
+    public static final String PANTS_CALF_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants4_calf.svg";
+    public static final String PANTS_BOTTOM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants5_bottom.svg";
+    public static final String PANTS_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pant6_len.svg";
+    public static final String PANTS_FLY_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants7_fly.svg";
+    public static final String PANTS_IN_SEAM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants8_Inseam.svg";
+    public static final String PANTS_CROTCH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Pants/pants9_Crotch.svg";
+    public static final String KURTA_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown.svg";
+    public static final String KURTA_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown1_len.svg";
+    public static final String KURTA_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown2_shoulder.svg";
+    public static final String KURTA_UPPER_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown3_upperchest.svg";
+    public static final String KURTA_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown4_bust.svg";
+    public static final String KURTA_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown5_waist.svg";
+    public static final String KURTA_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown6_seat.svg";
+    public static final String KURTA_ARMHOLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown7_armhole.svg";
+    public static final String KURTA_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown8_slevlen.svg";
+    public static final String KURTA_SLEEVE_CIRCUM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown9_slevcircum.svg";
+    public static final String KURTA_FRONT_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown10_frontneckdept.svg";
+    public static final String KURTA_BACK_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/menskurta/nightgown11_backneckdept.svg";
+    public static final String PYJAMA_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/Pajama/pajama.svg";
+    public static final String PYJAMA_HIP_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/Pajama/pajama_hip.svg";
+    public static final String PYJAMA_KNEE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/Pajama/pajama_knee.svg";
+    public static final String PYJAMA_ANKLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/Pajama/pajama_ankle.svg";
+    public static final String PYJAMA_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Kurta+pajama/Pajama/pajama_len.svg";
+    public static final String MENS_SUIT_UPPER_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt.svg";
+    public static final String MENS_SUIT_UPPER_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_length.svg";
+    public static final String MENS_SUIT_NECK_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_neck.svg";
+    public static final String MENS_SUIT_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_shoulder.svg";
+    public static final String MENS_SUIT_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_chest.svg";
+    public static final String MENS_SUIT_UPPER_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_waist.svg";
+    public static final String MENS_SUIT_UPPER_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_seat.svg";
+    public static final String MENS_SUIT_SLEEVE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_sleeve.svg";
+    public static final String MENS_SUIT_SLEEVE_CIRCUM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/Mens'suit(top)/shirt_sleevecircum.svg";
+    public static final String MENS_SUIT_LOWER_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants.svg";
+    public static final String MENS_SUIT_LOWER_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_waist.svg";
+    public static final String MENS_SUIT_LOWER_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_seat.svg";
+    public static final String MENS_SUIT_CALF_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_calf.svg";
+    public static final String MENS_SUIT_LOWER_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_length.svg";
+    public static final String MENS_SUIT_BOTTOM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_bottom.svg";
+    public static final String MENS_SUIT_FLY_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Mens+Suit/mens'suit(pants)/pants_fly.svg";
+    public static final String LADIES_SUIT_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/nightgown.svg";
+    public static final String LADIES_SUIT_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_length.svg";
+    public static final String LADIES_SUIT_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_shoulder.svg";
+    public static final String LADIES_SUIT_UPPER_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_upperchest.svg";
+    public static final String LADIES_SUIT_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_bust.svg";
+    public static final String LADIES_SUIT_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_waist.svg";
+    public static final String LADIES_SUIT_SEAT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NightGown/ng_seat.svg";
+    public static final String LADIES_SUIT_ARMHOLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_armhole.svg";
+    public static final String LADIES_SUIT_SLEEVE_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_slevelen.svg";
+    public static final String LADIES_SUIT_SLEEVE_CIRCUM_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_slevcircum.svg";
+    public static final String LADIES_SUIT_FRONT_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_frontneck.svg";
+    public static final String LADIES_SUIT_BACK_NECK_DEPTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/kameze/ng_backneck.svg";
+    public static final String LADIES_SUIT_LOWER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/salwar/pajama.svg";
+    public static final String LADIES_SUIT_SALWAR_HIP_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/salwar/pajama_hip.svg";
+    public static final String LADIES_SUIT_KNEE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/salwar/pajama_knee.svg";
+    public static final String LADIES_SUIT_ANKLE_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/salwar/pajama_ankle.svg";
+    public static final String LADIES_SUIT_SALWAR_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Salwar+kameez/salwar/pajama_len.svg";
+
+    public static final String NEHRU_JACKET_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NehruJacket/Nehrujacket.svg";
+
+    public static final String NEHRU_JACKET_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NehruJacket/nehrujacket1_length.svg";
+    public static final String NEHRU_JACKET_NECK_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NehruJacket/nehrujacket2_neck.svg";
+    public static final String NEHRU_JACKET_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NehruJacket/nehrujacket3_chest.svg";
+    public static final String NEHRU_JACKET_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/NehruJacket/nehrujacket4_waist.svg";
+
+    public static final String WAIST_COAT_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/WaistCoat/Waistcoat.svg";
+
+    public static final String WAIST_COAT_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/WaistCoat/Waistcoat1_length.svg";
+    public static final String WAIST_COAT_CHEST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/WaistCoat/Waistcoat2_chest.svg";
+    public static final String WAIST_COAT_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/WaistCoat/Waistcoat3_waist.svg";
+    public static final String PARDI_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper.svg";
+    public static final String RIDA_ABOVE_HEAD_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_above_head.svg";
+    public static final String RIDA_PARDI_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_shoulder.svg";
+    public static final String RIDA_AROUND_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_around_shoulder.svg";
+    public static final String RIDA_PARDI_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_length.svg";
+    public static final String RIDA_PARDI_GHER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_gher.svg";
+    public static final String RIDA_KAS_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/pardi/bhurkaupper_kas.svg";
+
+    public static final String LENGA_OUTFIT_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka.svg";
+    public static final String RIDA_LENGA_SHOULDER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka1_shoulder.svg";
+    public static final String RIDA_BUST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka2_Bust.svg";
+    public static final String RIDA_WAIST_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka3_waist.svg";
+    public static final String RIDA_HIPS_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka4_Hips.svg";
+    public static final String RIDA_LENGA_LENGTH_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka5_Length.svg";
+    public static final String RIDA_LENGA_GHER_IMAGE_LINK = "https://s3.amazonaws.com/" + staticBucketName + "/Measurement/Rida/lenga/bhurka6_gher.svg";
+
+}
+
