+++ src/main/java/com/darzee/shankh/constants/MeasurementKeys.java
@@ -1,0 +1,64 @@
+package com.darzee.shankh.constants;
+
+public class MeasurementKeys {
+    public static final String LENGTH_MEASUREMENT_KEY = "length";
+    public static final String WAIST_MEASUREMENT_KEY = "waist";
+    public static final String BLOUSE_LENGTH_MEASUREMENT_KEY = "blouse_length";
+
+    public static final String BUST_MEASUREMENT_KEY = "bust";
+    public static final String UPPER_CHEST_MEASUREMENT_KEY = "upper_chest";
+    public static final String BELOW_BUST_MEASUREMENT_KEY = "below_bust";
+    public static final String SHOULDER_MEASUREMENT_KEY = "shoulder";
+
+    public static final String ARM_HOLE_MEASUREMENT_KEY = "arm_hole";
+
+    public static final String SLEEVE_LENGTH_MEASUREMENT_KEY = "sleeve_length";
+    public static final String SLEEVE_CIRCUMFERENCE_MEASUREMENT_KEY = "sleeve_circumference";
+    public static final String BICEP_MEASUREMENT_KEY = "bicep";
+    public static final String ELBOW_ROUND_MEASUREMENT_KEY = "elbow_round";
+    public static final String CROSS_FRONT_MEASUREMENT_KEY = "cross_front";
+    public static final String CROSS_BACK_MEASUREMENT_KEY = "cross_back";
+
+    public static final String FRONT_NECK_DEPTH_MEASUREMENT_KEY = "front_neck_depth";
+
+    public static final String SHOULDER_TO_APEX_LENGTH_MEASUREMENT_KEY = "shoulder_to_apex_length";
+    public static final String APEX_TO_APEX_LENGTH_MEASUREMENT_KEY = "apex_to_apex_length";
+    public static final String BACK_NECK_DEPTH_MEASUREMENT_KEY = "back_neck_depth";
+    public static final String SEAT_MEASUREMENT_KEY = "seat";
+    public static final String GOWN_LENGTH_MEASUREMENT_KEY = "gown_length";
+    public static final String SHIRT_LENGTH_MEASUREMENT_KEY = "shirt_length";
+
+    public static final String NECK_MEASUREMENT_KEY = "neck";
+    public static final String CHEST_MEASUREMENT_KEY = "chest";
+    public static final String BOTTOM_WAIST_MEASUREMENT_KEY = "bottom_waist";
+    public static final String BOTTOM_SEAT_MEASUREMENT_KEY = "bottom_seat";
+    public static final String CALF_MEASUREMENT_KEY = "calf";
+
+    public static final String BOTTOM_MEASUREMENT_KEY = "bottom";
+    public static final String PANT_LENGTH_MEASUREMENT_KEY = "pant_length";
+    public static final String FLY_MEASUREMENT_KEY = "fly";
+    public static final String IN_SEAM_MEASUREMENT_KEY = "in_seam";
+    public static final String CROTCH_MEASUREMENT_KEY = "crotch";
+    public static final String KURTA_LENGTH_MEASUREMENT_KEY = "kurta_length";
+    public static final String PYJAMA_LENGTH_MEASUREMENT_KEY = "pyjama_length";
+    public static final String PYJAMA_HIP_MEASUREMENT_KEY = "pyjama_hip";
+
+    public static final String KNEE_MEASUREMENT_KEY = "knee";
+
+    public static final String ANKLE_MEASUREMENT_KEY = "ankle";
+    public static final String KAMEEZ_LENGTH_MEASUREMENT_KEY = "kameez_length";
+
+    public static final String SALWAR_HIP_MEASUREMENT_KEY = "salwar_hip";
+    public static final String SALWAR_LENGTH_MEASUREMENT_KEY = "salwar_length";
+    public static final String ABOVE_HEAD_MEASUREMENT_KEY = "above_head";
+    public static final String AROUND_SHOULDER_MEASUREMENT_KEY = "around_shoulder";
+    public static final String PARDI_SHOULDER_MEASUREMENT_KEY = "pardi_shoulder";
+    public static final String LENGA_SHOULDER_MEASUREMENT_KEY = "lenga_shoulder";
+    public static final String PARDI_LENGTH_MEASUREMENT_KEY = "pardi_length";
+    public static final String LENGA_LENGTH_MEASUREMENT_KEY = "lenga_length";
+    public static final String PARDI_GHER_MEASUREMENT_KEY = "pardi_gher";
+    public static final String LENGA_GHER_MEASUREMENT_KEY = "lenga_gher";
+    public static final String KAS_MEASUREMENT_KEY = "kas";
+    public static final String HIP_MEASUREMENT_KEY = "hip";
+
+}
