+++ src/main/java/com/darzee/shankh/constants/MeasurementTitles.java
@@ -1,0 +1,164 @@
+package com.darzee.shankh.constants;
+
+public class MeasurementTitles {
+
+    public static final String DRESS_LENGTH_TITLE = "Length";
+    public static final String DRESS_SHOULDER_TITLE = "Shoulder";
+    public static final String DRESS_UPPER_CHEST_TITLE = "Upper Chest";
+    public static final String DRESS_BUST_TITLE = "Bust";
+    public static final String DRESS_WAIST_TITLE = "Waist";
+    public static final String DRESS_SEAT_TITLE = "Seat";
+    public static final String DRESS_ARMHOLE_TITLE = "Arm Hole";
+    public static final String DRESS_SLEEVE_LENGTH_TITLE = "Sleeve Length";
+    public static final String DRESS_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+
+    public static final String DRESS_FRONT_NECK_DEPTH_TITLE = "Front Neck Depth";
+    public static final String DRESS_BACK_NECK_DEPTH_TITLE = "Back Neck Depth";
+
+    public static final String GOWN_LENGTH_TITLE = "Gown Length";
+
+    public static final String GOWN_SEAT_TITLE = "Seat";
+    public static final String GOWN_SHOULDER_TITLE = "Shoulder";
+    public static final String GOWN_UPPER_CHEST_TITLE = "Upper Chest";
+    public static final String GOWN_BUST_TITLE = "Bust";
+    public static final String GOWN_WAIST_TITLE = "Waist";
+    public static final String GOWN_ARMHOLE_TITLE = "Armhole";
+    public static final String GOWN_SLEEVE_LENGTH_TITLE = "Sleeve Length";
+    public static final String GOWN_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+    public static final String GOWN_FRONT_NECK_TITLE = "Front Neck Depth";
+    public static final String GOWN_BACK_NECK_DEPTH_TITLE = "Back Neck Depth";
+
+    public static final String UNDER_SKIRT_WAIST_TITLE = "Waist";
+
+    public static final String UNDER_SKIRT_LENGTH_TITLE = "Length";
+
+    public static final String BLOUSE_LENGTH_TITLE = "Blouse Length";
+    public static final String BLOUSE_BUST_TITLE = "Bust";
+    public static final String BLOUSE_UPPER_CHEST_TITLE = "Upper Chest";
+    public static final String BLOUSE_BELOW_BUST_TITLE = "Below Bust";
+    public static final String BLOUSE_SHOULDER_TITLE = "Shoulder";
+    public static final String BLOUSE_ARMHOLE_TITLE = "Armhole";
+    public static final String BLOUSE_SLEEVE_LENGTH_TITLE = "Sleeve Length";
+    public static final String BLOUSE_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+    public static final String BLOUSE_BICEP_TITLE = "Bicep";
+    public static final String BLOUSE_ELBOW_ROUND_TITLE = "Elbow Round";
+    public static final String BLOUSE_FRONT_NECK_DEPTH_TITLE = "Front Neck Depth";
+    public static final String BLOUSE_SHOULDER_TO_APEX_TITLE = "Shoulder to Apex";
+    public static final String BLOUSE_APEX_TO_APEX_TITLE = "Apex to Apex";
+    public static final String BLOUSE_BACK_NECK_DEPTH_TITLE = "Back Neck Depth";
+    public static final String BLOUSE_CROSS_FRONT_TITLE = "Cross Front";
+    public static final String BLOUSE_CROSS_BACK_TITLE = "Cross Back";
+    public static final String SHIRT_LENGTH_TITLE = "Length";
+    public static final String SHIRT_NECK_TITLE = "Neck";
+    public static final String SHIRT_SHOULDER_TITLE = "Shoulder";
+    public static final String SHIRT_CHEST_TITLE = "Chest";
+    public static final String SHIRT_WAIST_TITLE = "Waist";
+    public static final String SHIRT_SEAT_TITLE = "Seat";
+    public static final String SHIRT_SLEEVE_LENGTH_TITLE = "Sleeve";
+    public static final String SHIRT_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+
+    public static final String PANTS_WAIST_TITLE = "Waist";
+
+    public static final String PANTS_SEAT_TITLE = "Seat";
+
+    public static final String PANTS_THIGH_TITLE = "Thigh Circum";
+
+    public static final String PANTS_CALF_TITLE = "Calf/Knee";
+
+    public static final String PANTS_BOTTOM_TITLE = "Bottom/Bells";
+
+    public static final String PANTS_LENGTH_TITLE = "Length";
+
+    public static final String PANTS_FLY_TITLE = "Fly(Ply)";
+    public static final String PANTS_IN_SEAM_TITLE = "In-Seam";
+    public static final String PANTS_CROTCH_TITLE = "Crotch";
+    public static final String KURTA_LENGTH_TITLE = "Kurta Length";
+    public static final String KURTA_SHOULDER_TITLE = "Shoulder";
+    public static final String KURTA_UPPER_CHEST_TITLE = "Upper Chest";
+    public static final String KURTA_BUST_TITLE = "Bust";
+    public static final String KURTA_WAIST_TITLE = "Waist";
+    public static final String KURTA_SEAT_TITLE = "Seat";
+    public static final String KURTA_ARMHOLE_TITLE = "Armhole";
+    public static final String KURTA_SLEEVE_LENGTH_TITLE = "Sleeve Length";
+    public static final String KURTA_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+    public static final String KURTA_FRONT_NECK_DEPTH_TITLE = "Front Neck Depth";
+    public static final String KURTA_BACK_NECK_DEPTH_TITLE = "Back Neck Depth";
+    public static final String PYJAMA_HIP_TITLE = "Pajama Hip";
+    public static final String PYJAMA_KNEE_TITLE = "Knee";
+    public static final String PYJAMA_ANKLE_TITLE = "Ankle";
+    public static final String PYJAMA_LENGTH_TITLE = "Pajama Length";
+
+    public static final String MENS_SUIT_UPPER_LENGTH_TITLE = "Length";
+    public static final String MENS_SUIT_NECK_TITLE = "Neck";
+    public static final String MENS_SUIT_SHOULDER_TITLE = "Shoulder";
+    public static final String MENS_SUIT_CHEST_TITLE = "Chest";
+    public static final String MENS_SUIT_UPPER_WAIST_TITLE = "Waist";
+    public static final String MENS_SUIT_UPPER_SEAT_TITLE = "Seat";
+    public static final String MENS_SUIT_SLEEVE_TITLE = "Sleeve";
+    public static final String MENS_SUIT_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+    public static final String MENS_SUIT_LOWER_WAIST_TITLE = "Waist";
+    public static final String MENS_SUIT_LOWER_SEAT_TITLE = "Seat";
+    public static final String MENS_SUIT_CALF_TITLE = "Calf/Knee";
+    public static final String MENS_SUIT_BOTTOM_TITLE = "Bottom/Bells";
+    public static final String MENS_SUIT_LOWER_LENGTH_TITLE = "Length";
+    public static final String MENS_SUIT_FLY_TITLE = "Fly(Ply)";
+    public static final String LADIES_SUIT_KAMEEZ_LENGTH_TITLE = "Kameez Length";
+    public static final String LADIES_SUIT_SHOULDER_TITLE = "Shoulder";
+    public static final String LADIES_SUIT_UPPER_CHEST_TITLE = "Upper Chest";
+    public static final String LADIES_SUIT_BUST_TITLE = "Bust";
+    public static final String LADIES_SUIT_WAIST_TITLE = "Waist";
+    public static final String LADIES_SUIT_SEAT_TITLE = "Seat";
+    public static final String LADIES_SUIT_ARMHOLE_TITLE = "Armhole";
+    public static final String LADIES_SUIT_SLEEVE_LENGTH_TITLE = "Sleeve Length";
+    public static final String LADIES_SUIT_SLEEVE_CIRCUM_TITLE = "Sleeve Circum";
+    public static final String LADIES_SUIT_FRONT_NECK_DEPTH_TITLE = "Front Neck Depth";
+    public static final String LADIES_SUIT_BACK_NECK_DEPTH_TITLE = "Back Neck Depth";
+    public static final String LADIES_SUIT_SALWAR_HIP_TITLE = "Salwar Hip";
+    public static final String LADIES_SUIT_KNEE_TITLE = "Knee";
+    public static final String LADIES_SUIT_ANKLE_TITLE = "Ankle";
+    public static final String LADIES_SUIT_SALWAR_LENGTH_TITLE = "Salwar Length";
+
+    public static final String NEHRU_JACKET_LENGTH_TITLE = "Length";
+
+    public static final String NEHRU_JACKET_NECK_TITLE = "Neck";
+
+    public static final String NEHRU_JACKET_CHEST_TITLE = "Chest";
+
+    public static final String NEHRU_JACKET_WAIST_TITLE = "Waist";
+    public static final String WAIST_COAT_LENGTH_TITLE = "Length";
+    public static final String WAIST_COAT_CHEST_TITLE = "Chest";
+
+    public static final String WAIST_COAT_WAIST_TITLE = "Waist";
+
+    public static final String RIDA_ABOVE_HEAD_TITLE = "Above Head";
+    public static final String RIDA_PARDI_SHOULDER_TITLE = "Shoulder";
+    public static final String RIDA_AROUND_SHOULDER_TITLE = "Around Shoulder";
+    public static final String RIDA_PARDI_LENGTH_TITLE = "Length";
+    public static final String RIDA_PARDI_GHER_TITLE = "Gher";
+    public static final String RIDA_KAS_TITLE = "Kas";
+    public static final String RIDA_LENGA_SHOULDER_TITLE = "Shoulder";
+    public static final String RIDA_BUST_TITLE = "Bust";
+    public static final String RIDA_WAIST_TITLE = "Waist";
+    public static final String RIDA_HIP_TITLE = "Hips";
+    public static final String RIDA_LENGA_LENGTH_TITLE = "Length";
+    public static final String RIDA_LENGA_GHER_TITLE = "Gher";
+
+    public static final String DRESS_OUTFIT_TYPE_HEADING = "Dress";
+    public static final String PANTS_OUTFIT_TYPE_HEADING = "Pants";
+    public static final String SHIRT_OUTFIT_TYPE_HEADING = "Shirt";
+    public static final String NIGHT_GOWN_OUTFIT_TYPE_HEADING = "Night Gown";
+    public static final String UNDER_SHIRT_OUTFIT_TYPE_HEADING = "Underskirt";
+    public static final String SAREE_BLOUSE_OUTFIT_TYPE_HEADING = "Saree Blouse";
+    public static final String MENS_KURTA_OUTFIT_TYPE_HEADING = "Kurta Pajama (Top)";
+    public static final String MENS_PYJAMA_OUTFIT_TYPE_HEADING = "Kurta Pajama (Bottom)";
+    public static final String MENS_SUIT_TOP_OUTFIT_TYPE_HEADING = "Men's Suit (Top)";
+    public static final String MENS_SUIT_PANTS_OUTFIT_TYPE_HEADING = "Men's Suit (Bottom)";
+    public static final String LADIES_SUIT_KAMEEZ_OUTFIT_TYPE_HEADING = "Ladies Suit (Top)";
+    public static final String LADIES_SUIT_SALWAR_OUTFIT_TYPE_HEADING = "Ladies Suit (Bottom)";
+    public static final String NEHRU_JACKET_OUTFIT_TYPE_HEADING = "Nehru Jacket";
+    public static final String WAIST_COAT_OUTFIT_TYPE_HEADING = "Waist Coat";
+    public static final String RIDA_PARDI_OUTFIT_TYPE_HEADING = "Pardi";
+    public static final String RIDA_LENGA_OUTFIT_TYPE_HEADING = "Lenga";
+
+
+}
