+++ src/main/resources/application-dev.properties
@@ -8,3 +8,4 @@ amazonProperties.s3.endpointUrl=https://s3.amazonaws.com
 amazonProperties.s3.accessKey=AKIA5TC6Q7Z7Q25LCETU
 amazonProperties.s3.secretKey=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
 amazonProperties.s3.bucketName=darzee.backend.stage
+amazonProperties.s3.staticBucketName=darzee.backend.stage.static
